Pom.xml and properties.log completed
In order to use all the functions of Mockito and Hamcrest, I had to
import their dependencies.

The properties.log file was incomplete, it was only registering the
loggers on the console, they are now also being added to the
log4j2.properties file => available under ressources. (it is advised to
have such a structure when we are in the debugging/creating an app).