30 Minutes free Parking
The users are now able to park for 30 minutes for free. I first created
the unit test, testing that if somebody parks for 25 minutes the price
would be of 0. It ended of course in an error and then adapted my code
to create this feature. The Unit test are now passing normally. Yayyy!
haha please tell me if somebody actually read that! it's 9:31 and I
enjoy being productive! Stay classy San Fransisco!