Story #2: Method in FareCalculatorService has been written
PIECE OF CAKE! The method for the 5% discount for returning users has
been written, it works perfectly and the unit test is passing! all good!

I now have to actually change the code so that we recognise returning
visitors and ad the greeting text for them! I guess that includes
integration testing! let's do that Mofo!!!