Story #2 The method and testing are ready, next step: implementation!
We now have a method that checks if a user was already once parked in
AND out of the parking. Nobody can then use the registration number from
the car in front of them to get the 5% on first and last visit.
Hahahaha not that anybody would think about that. Let's stay on the sure
side of the fence though ;)

Ready for implementation now