Test if an outTime is set in the DB
I noticed that the test was failing 30% of the time. When it was the
case, alert I got was saying that the outTime was invalid. Looking
closer I notice that the inTime and outTime were sometimes the same when
the failure occurred. That make sense since the code is run on a
millisecond base and the date on a second basis.

My solution: I created the Calendar variable "laterOutTime" that is set
to the instant it's called upon and added one second to this instant.
inTime and outTime will now always have at least one second of
difference.

The best part, the Users will have to pay this second and will never
notice the difference => Like a true capitalists boss. find more
@GoldmansaxTradingTechniquesUsingAlgorithmsToScrewHumanTraders