Story #2 Last Changes
That was a looot of work! I thought I could implement Story number two
two without having to touch to the database. I was wrong!

I added a column to the database called RETURNING_USER, I then had to
add some setters and getter for this column. After that some changes to
the InputReaderUtil were necessary since I had the isReturningUser
needed now to update the DB and not only change a variable. This was
followed by the implementation of the code in the FareCalculator
Service.

OF course some test were not passing anymore. So I finally had to adapt
those tests to the newly updated methods! but here we are! hahahha all
done!

Next phase: Jacoco, Surefire & Bug Reports!!