Story #2 Last last Changes
hahaha that took longer than expected. Adding one line in the database
in order to know if a user is a returning User had a lot of consequences
everywhere in the code.

TicketDAO: The saveTicket and getTicket methods had to be updated,
adding one more line of code in each.

DBConstants: Since the TicketDAO methods were changed and one more line
was added, the SQL statements had to be set accordingly.

InputReaderUtil: I realised that the method "readifReturningUser" did
not belong here since this class is not a class that is made to
communicate with the Database. I deleted the method.

ParkingService: This is where the above mentioned method belongs! it is
now cleanly incorporated.

ParkingDataBaseIT: The tests had to be slightly changed to fit the newly
placed feature!

Voilà! nous y sommes mon amis. Ce fut un combat, mais nous sommes les
vainqueurs!