DAO Testing #3 Exceptions
I created a new class for the Junit tests. It was necessary in order to
be able to test the exceptions => the import required for the exception
tests were not compatible with the other DAO tests.

Classpath: I added the Junit Library to my class path

ParkingDataBaseIT & TicketDAO.java: minor changes, spacing etc...